improvement(#52): Add dashboard page
