+++ frontend/src/utils/constants.ts
@@ -1,4 +1,4 @@
 export const BACKEND_API_URL =
   process.env.BACKEND_API_URL || "http://localhost:42069";
 export const SIGN_OUT_CALLBACK_URL = "/sign-in";
-export const SIGN_IN_CALLBACK_URL = "/";
+export const SIGN_IN_CALLBACK_URL = "/dashboard";
